backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.40.0
google-generativeai>=0.8.0
gunicorn>=22.0.0
pypdf>=5.0.0
python-docx>=1.1.0
qdrant-client>=1.12.0
groq>=0.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
